requirements.txt
garminconnect==0.2.19
garth==0.4.46
google_api_python_client==2.147.0
protobuf==5.28.2
python-dotenv==1.0.1
Requests==2.32.3
logger==1.4
